requirements.txt

altgraph==0.17.3
anyio==4.3.0
asttokens==2.4.1
asyncio==3.4.3
attrs==24.2.0
autolab-core==1.1.1
autopep8==2.0.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
et-xmlfile==1.1.0
executing==2.1.0
EZSheets==2023.3.14
fastjsonschema==2.20.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
flask-mysql-connector==1.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
freetype-py==2.5.1
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.21.0
httpx==0.27.0
idna==3.4
imageio==2.35.1
ipykernel==6.29.5
ipython==8.27.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
itsdangerous==2.1.2
jdcal==1.4.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lazy_loader==0.4
lxml==4.9.2
manifold3d==2.5.1
mapbox_earcut==1.0.2
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
multiprocess==0.70.16
mysql-connector-python==9.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.2
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prettytable==3.11.0
prompt_toolkit==3.0.48
protobuf==4.22.1
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycollada==0.8
pycryptodome==3.17
pyglet==2.0.17
Pygments==2.18.0
PyInputPlus==0.2.12
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyMySQL==1.1.1
PyOpenGL==3.1.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyreadline==2.1
pyrender==0.1.45
PySimpleValidate==0.2.12
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.28.2
requests-oauthlib==1.3.1
rpds-py==0.20.0
rsa==4.9
Rtree==1.3.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sql==2022.4.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
stdiomask==0.0.6
svg.path==6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
trimesh==4.4.9
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.15
vhacdx==0.0.8.post1
visualization==1.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
xatlas==0.0.9
xxhash==3.5.0


